Dual compile Windows RSA into Algorithms
