+++ src/Common/src/Internal/Cryptography/CngCommon.Hash.cs
@@ -48,4 +48,3 @@ namespace Internal.Cryptography
         }
     }
 }
-
