+++ src/Common/src/Interop/Windows/BCrypt/Interop.AsymmetricEncryption.Types.cs
