+++ src/Common/src/Interop/Windows/NCrypt/Interop.NCryptFreeObject.cs
@@ -1,0 +1,15 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.Runtime.InteropServices;
+
+internal static partial class Interop
+{
+    internal static partial class NCrypt
+    {
+        [DllImport(Interop.Libraries.NCrypt, CharSet = CharSet.Unicode)]
+        internal static extern ErrorCode NCryptFreeObject(IntPtr hObject);
+    }
+}
