+++ src/Common/src/System/Security/Cryptography/RSACng.cs
@@ -1,0 +1,62 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.IO;
+using Internal.Cryptography;
+
+namespace System.Security.Cryptography
+{
+#if INTERNAL_ASYMMETRIC_IMPLEMENTATIONS
+    internal static partial class RSAImplementation
+    {
+#endif
+    public sealed partial class RSACng : RSA
+    {
+        /// <summary>
+        ///     Create an RSACng algorithm with a random 2048 bit key pair.
+        /// </summary>
+        public RSACng()
+            : this(2048)
+        {
+        }
+
+        /// <summary>
+        ///     Creates a new RSACng object that will use a randomly generated key of the specified size.
+        ///     Valid key sizes range from 512 to 16384 bits, in increments of 64 bits. It is suggested that a
+        ///     minimum size of 2048 bits be used for all keys.
+        /// </summary>
+        /// <param name="keySize">Size of the key to generate, in bits.</param>
+        /// <exception cref="CryptographicException">if <paramref name="keySize" /> is not valid</exception>
+        public RSACng(int keySize)
+        {
+            KeySize = keySize;
+        }
+
+        public override KeySizes[] LegalKeySizes
+        {
+            get
+            {
+                // See https://msdn.microsoft.com/en-us/library/windows/desktop/bb931354(v=vs.85).aspx
+                return new KeySizes[]
+                {
+                    // All values are in bits.
+                    new KeySizes(minSize: 512, maxSize: 16384, skipSize: 64),
+                };
+            }
+        }
+
+        protected override byte[] HashData(byte[] data, int offset, int count, HashAlgorithmName hashAlgorithm)
+        {
+            return CngCommon.HashData(data, offset, count, hashAlgorithm);
+        }
+
+        protected override byte[] HashData(Stream data, HashAlgorithmName hashAlgorithm)
+        {
+            return CngCommon.HashData(data, hashAlgorithm);
+        }
+    }
+#if INTERNAL_ASYMMETRIC_IMPLEMENTATIONS
+    }
+#endif
+}
