+++ src/System.Security.Cryptography.Algorithms/src/project.json
@@ -7,6 +7,7 @@
         "System.IO": "4.0.10",
         "System.Resources.ResourceManager": "4.0.0",
         "System.Runtime": "4.0.20",
+        "System.Runtime.Extensions": "4.0.0",
         "System.Runtime.InteropServices": "4.0.20",
         "System.Security.Cryptography.Primitives": "4.0.0-rc3-23808",
         "System.Text.Encoding.Extensions": "4.0.0"
