--- src/System.Security.Cryptography.Cng/src/System/Security/Cryptography/RSACng.HashData.cs
@@ -1,26 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System;
-using System.Diagnostics;
-using System.IO;
-
-using Internal.Cryptography;
-
-namespace System.Security.Cryptography
-{
-    public sealed partial class RSACng : RSA
-    {
-        protected override byte[] HashData(byte[] data, int offset, int count, HashAlgorithmName hashAlgorithm)
-        {
-            return CngCommon.HashData(data, offset, count, hashAlgorithm);
-        }
-
-        protected override byte[] HashData(Stream data, HashAlgorithmName hashAlgorithm)
-        {
-            return CngCommon.HashData(data, hashAlgorithm);
-        }
-    }
-}
-
