+++ src/System.Security.Cryptography.Cng/src/System/Security/Cryptography/RSACng.cs
@@ -12,26 +12,6 @@ namespace System.Security.Cryptography
     public sealed partial class RSACng : RSA
     {
         /// <summary>
-        ///     Create an RSACng algorithm with a random 2048 bit key pair.
-        /// </summary>
-        public RSACng()
-            : this(2048)
-        {
-        }
-
-        /// <summary>
-        ///     Creates a new RSACng object that will use a randomly generated key of the specified size.
-        ///     Valid key sizes range from 512 to 16384 bits, in increments of 64 bits. It is suggested that a
-        ///     minimum size of 2048 bits be used for all keys.
-        /// </summary>
-        /// <param name="keySize">Size of the key to generate, in bits.</param>
-        /// <exception cref="CryptographicException">if <paramref name="keySize" /> is not valid</exception>
-        public RSACng(int keySize)
-        {
-            KeySize = keySize;
-        }
-
-        /// <summary>
         ///     Creates a new RSACng object that will use the specified key. The key's
         ///     <see cref="CngKey.AlgorithmGroup" /> must be Rsa. This constructor
         ///     creates a copy of the key. Hence, the caller can safely dispose of the 
@@ -51,19 +31,6 @@ namespace System.Security.Cryptography
             Key = CngAlgorithmCore.Duplicate(key);
         }
 
-        public override KeySizes[] LegalKeySizes
-        {
-            get
-            {
-                // See https://msdn.microsoft.com/en-us/library/windows/desktop/bb931354(v=vs.85).aspx
-                return new KeySizes[] 
-                {
-                    // All values are in bits.
-                    new KeySizes(minSize: 512, maxSize: 16384, skipSize: 64), 
-                }; 
-            }
-        }
-
         protected override void Dispose(bool disposing)
         {
             _core.Dispose();
@@ -72,4 +39,3 @@ namespace System.Security.Cryptography
         private CngAlgorithmCore _core;
     }
 }
-
